fix(seed): row_number() is bigint, and date - bigint has no operator
`rn % 7` inherits bigint from row_number(), so the shift arithmetic failed
with "operator does not exist: date - bigint". PostgreSQL defines
date - integer, not date - bigint.

Third type assumption this schema has caught. The pattern is the same each
time: inferring a type from what a value means rather than from what the
database declares. A row number is conceptually small; it is not an int.

+++ supabase/seed/20260810_realign_open_deal_activity.sql
@@ -111,12 +111,14 @@ ranked as (
   from open_deals
 ),
 targets as (
+  -- ::int is load-bearing. row_number() is bigint, so `rn % 7` is bigint, and
+  -- `date - bigint` has no operator in PostgreSQL — only `date - integer`.
   select id, last_act,
-         case
-           when rn <= 22 then (rn % 7)              -- 0-6  days silent
-           when rn <= 27 then 7 + (rn % 8)          -- 7-14 days
-           else               15 + (rn % 21)        -- 15-35 days
-         end as target_days
+         (case
+            when rn <= 22 then (rn % 7)             -- 0-6  days silent
+            when rn <= 27 then 7 + (rn % 8)         -- 7-14 days
+            else               15 + (rn % 21)       -- 15-35 days
+          end)::int as target_days
   from ranked
 ),
 shifts as (
